Client refactor (#6)
* feat: rocket master test

* fix: downgraded react version

* feat: new database

* feat: changed ids, new tables, stats route

* feat: collector implementation for users folder

* wip: added typescript

* feat: trade with new DB, fixed code

* feat: changed all components and pages to typescript

* feat: pack new db, some bugfixes

* fix: route collision, notifications one route

* feat: notifications deleteall new database

* feat: added auth provider

* feat: added useAuth hook

* feat: added private routes

* fix: fixed collector_exists typo

* feat: create collector

* feat: collector favorite add

* feat: collector favorite remove

* fix: fixed malformated sql statement

* feat: added refresh token login

* feat: added logout

* fix: refresh token now gets correct expiry date

* feat: added refresh token rotation

* fix: fixed settings not showing when loggedin

* feat: added remember me button

* fix: removed debug button

* fix: added tutorial card

* feat: added new navigation

* feat: upgraded to react 17

* feat: upgraded react-router-dom to 6.3.0

* fix: cleaned up backend code

* feat: Id implementation

* feat: id serde

* wip: some restructure

* feat: added collector list

* wip: cleanup

* feat: progress

* feat: progress

* feat: progress

* feat: progress

* feat: progress

* feat: progress

* feat: progress

* feat: finished refactor

+++ server/src/notifications/deleteall/sql.rs
@@ -1,18 +1,24 @@
 use sqlx::mysql::MySqlQueryResult;
 
+
 use crate::sql::Sql;
 use crate::shared::Id;
 
-pub async fn delete_all_notifications(sql: &Sql, user_id: Id) -> Result<u64, sqlx::Error> {
+pub async fn delete_all_notifications(sql: &Sql, user_id: &Id, collector_id: &Option<Id>) -> Result<u64, sqlx::Error> {
     let mut con = sql.get_con().await?;
+    let sql = format!("DELETE FROM notifications WHERE uid = ? {};",match collector_id{ 
+        None => "",
+        Some(_collector_id) => "AND coid = ?"
+    });
 
-    let result: MySqlQueryResult = sqlx::query(
-        "DELETE FROM notifications
-         WHERE
-         uid = ?;")
-        .bind(user_id)
-        .execute(&mut con)
-        .await?;
+
+    let mut stmt = sqlx::query(&sql).bind(user_id);
+
+    if let Some(collector_id) = collector_id{
+       stmt = stmt.bind(collector_id);  
+    }
+
+    let result: MySqlQueryResult = stmt.execute(&mut con).await?;
 
     Ok(result.rows_affected())
 }
